source-code/dvpa/requirements.txt
appnope==0.1.2
backcall==0.2.0
click==6.6
decorator==4.4.2
Flask==0.11.1
flask-mongoengine==0.8
Flask-MySQL==1.5.1
Flask-MySQLdb==0.2.0
flask-peewee==0.6.7
Flask-Script==2.0.5
Flask-WTF==0.12
ipdb==0.13.4
ipython==7.19.0
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.17.2
Jinja2==2.8
MarkupSafe==0.23
mongoengine==0.10.6
mysqlclient==2.0.1
parso==0.7.1
peewee==2.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.8
ptyprocess==0.6.0
Pygments==2.7.2
pymongo==3.3.0
PyMySQL==0.10.1
PyYAML==5.1
traitlets==5.0.5
wcwidth==0.2.5
Werkzeug==0.11.11
wtf-peewee==0.2.6
WTForms==2.1